+++ ios/Classes/pytorch_ffi.cpp
@@ -121,6 +121,78 @@ model_inference(int index,float *input_data_ptr,int input_length) {
     // Return the output data structure
     return output;
 }
+cv::Mat decode_image(unsigned char* data, int input_length) {
+    cv::_InputArray inputArray(data, input_length);
+    return cv::imdecode(inputArray, cv::IMREAD_COLOR);
+}
+cv::Mat preprocess_mat(cv::Mat img, int height, int width, float* mean_values, float* std_values) {
+    cv::Mat imgRGB;
+    cv::cvtColor(img, imgRGB, cv::COLOR_BGR2RGB);
+
+    cv::Size sizeDesired(width, height);
+    cv::Mat imgResized;
+    cv::resize(imgRGB, imgResized, sizeDesired);
+
+    imgResized.convertTo(imgResized, CV_32FC3, 1.0f / 255.0f);
+
+    cv::Mat mean(1, 1, CV_32FC3, cv::Scalar(mean_values[0], mean_values[1], mean_values[2]));
+    cv::Mat std(1, 1, CV_32FC3, cv::Scalar(std_values[0], std_values[1], std_values[2]));
+
+    std::vector<cv::Mat> channels;
+    cv::split(imgResized, channels);
+
+    for (int i = 0; i < 3; ++i) {
+        cv::Mat mean_ch = mean.at<cv::Vec3f>(0)[i] * cv::Mat::ones(imgResized.size(), CV_32F);
+        cv::Mat std_ch = std.at<cv::Vec3f>(0)[i] * cv::Mat::ones(imgResized.size(), CV_32F);
+
+        cv::subtract(channels[i], mean_ch, channels[i]);
+        cv::divide(channels[i], std_ch, channels[i]); 
+    }
+
+    cv::Mat imgNormalized;
+    cv::merge(channels, imgNormalized);
+    return imgNormalized;
+}
+OutputData run_inference(int index, cv::Mat imgNormalized, int height, int width, int objectDetectionFlag, float* output_data) {
+    OutputData output;
+    try {
+        torch::jit::Module model = models.at(index);
+        torch::Tensor tensor_image = torch::from_blob(imgNormalized.data, {1, height, width, 3}, torch::kFloat32);
+        tensor_image = tensor_image.permute({0, 3, 1, 2});
+
+        torch::Tensor output_tensor;
+        if (objectDetectionFlag == 1) {
+            torch::jit::IValue outputTuple = model.forward({tensor_image}).toTuple();
+            output_tensor = outputTuple.toTuple()->elements()[0].toTensor();
+        } else {
+            output_tensor = model.forward({tensor_image}).toTensor();
+        }
+
+        int tensor_length = output_tensor.numel();
+        memcpy(output_data, output_tensor.data_ptr<float>(), sizeof(float) * tensor_length);
+
+        output.values = output_data;
+        output.length = tensor_length;
+        output.exception = "";
+
+    } catch (const std::exception& e) {
+        std::string exceptionMessage = e.what();
+        output.exception = strdup(exceptionMessage.c_str());
+    }
+    return output;
+}
+void rotateMat(cv::Mat &matImage, int rotation) {
+    if (rotation == 90) {
+        transpose(matImage, matImage);
+        flip(matImage, matImage, 1); //transpose+flip(1)=CW
+    } else if (rotation == 270) {
+        transpose(matImage, matImage);
+        flip(matImage, matImage, 0); //transpose+flip(0)=CCW
+    } else if (rotation == 180) {
+        flip(matImage, matImage, -1);    //flip(-1)=180
+    }
+}
+
 
 /* Function: image_model_inference
  * Input: 
@@ -132,101 +204,26 @@ model_inference(int index,float *input_data_ptr,int input_length) {
  * Description: This function runs inference on the given image data and returns output data.
  * It also captures any exception that might occur during inference and records it in OutputData.
  */
+ extern "C" __attribute__((visibility("default"))) __attribute__((used)) OutputData
+
+ image_model_inference(int index, unsigned char* data, int input_length, int height, int width, int objectDetectionFlag, float* mean_values, float* std_values, float* output_data) {
+    cv::Mat img = decode_image(data, input_length);
+    cv::Mat imgNormalized = preprocess_mat(img, height, width, mean_values, std_values);
+    return run_inference(index, imgNormalized, height, width, objectDetectionFlag, output_data);
+}
+
 extern "C" __attribute__((visibility("default"))) __attribute__((used)) OutputData
-image_model_inference(int index, unsigned char* data,int input_length, int height, int width, int objectDetectionFlag, float* mean_values, float* std_values,float* output_data) {
-    // Define the output data structure
-    struct OutputData output;
-    try {
-        cv::_InputArray inputArray(data,input_length);
-        cv::Mat img = cv::imdecode(inputArray, cv::IMREAD_COLOR);
-        
-        cv::Mat imgRGB;
-        cv::cvtColor(img, imgRGB, cv::COLOR_BGR2RGB);
-        // Convert the received image data into an OpenCV Mat
 
-        // Use the provided height and width as the desired size
-        cv::Size sizeDesired(width, height);
-
-        // Resize the image to the desired size
-        cv::Mat imgResized;
-        cv::resize(imgRGB, imgResized, sizeDesired);
-
-        // Convert image to float and normalize to [0, 1]
-        cv::Mat imgFloat;
-        // convert [unsigned int] to [float]
-        imgResized.convertTo(imgFloat, CV_32FC3, 1.0f / 255.0f);
-
-        // Keep mean and std as 3-element vectors
-        cv::Mat mean(1, 1, CV_32FC3, cv::Scalar(mean_values[0], mean_values[1], mean_values[2]));
-        cv::Mat std(1, 1, CV_32FC3, cv::Scalar(std_values[0], std_values[1], std_values[2]));
-
-        // Split image channels 
-        std::vector<cv::Mat> channels;
-        cv::split(imgFloat, channels);
-
-        // Normalize each channel  
-        for (int i = 0; i < 3; ++i) {
-
-        cv::Mat mean_ch = mean.at<cv::Vec3f>(0)[i] * cv::Mat::ones(imgFloat.size(), CV_32F);
-        cv::Mat std_ch = std.at<cv::Vec3f>(0)[i] * cv::Mat::ones(imgFloat.size(), CV_32F);
-
-        cv::subtract(channels[i], mean_ch, channels[i]);
-        cv::divide(channels[i], std_ch, channels[i]); 
-
-        }
-
-        // Merge channels back
-        cv::Mat imgNormalized;
-        cv::merge(channels, imgNormalized);
-
-        // Load the PyTorch model
-        torch::jit::Module model = models.at(index);
-        // Convert the normalized image data into a PyTorch tensor
-        torch::Tensor tensor_image = torch::from_blob(imgNormalized.data, {1, height, width, 3}, torch::kFloat32);
-        tensor_image = tensor_image.permute({0, 3, 1, 2});
-
-
-        torch::Tensor output_tensor;
-        if (objectDetectionFlag==1) {
-            // Run the object detection model
-
-//            // Run the model with the input tensor and get the output tuple of tensors
-//            auto output_tuple_ptr = model.forward({tensor_image}).toTuple();
-//
-//            // Example: Extract the first tensor from the tuple
-//            auto& output_tuple = *output_tuple_ptr;
-//
-//            // Access the elements of the tuple using the appropriate methods
-//            output_tensor = output_tuple.elements()[0].toTensor();
-
-            torch::jit::IValue outputTuple = model.forward({tensor_image}).toTuple();
-            output_tensor = outputTuple.toTuple()->elements()[0].toTensor();
-
-        }
-        else {
-            // Run the model with the input tensor and get the output tensor
-            output_tensor = model.forward({tensor_image}).toTensor();
-
-        }
-        // Get the number of elements in the output tensor
-        int tensor_length = output_tensor.numel();
-        
-        // Allocate memory for output data and copy data from the output tensor
-        // float *output_data = static_cast<float*>(malloc(sizeof(float) * tensor_length));
-        memcpy(output_data, output_tensor.data_ptr<float>(), sizeof(float) * tensor_length);
-
-        // Store the output data and length in the output data structure
-        output.values = output_data;
-        output.length = tensor_length;
-        output.exception = "";  // Empty string indicates no exception occurred
+ camera_model_inference(int index, unsigned char* data,int rotation,int isYUV, int model_image_height, int model_image_width,int camera_image_height,int camera_image_width, int objectDetectionFlag, float* mean_values, float* std_values, float* output_data) {
+    cv::Mat img;
+    if (isYUV==1) {
+        cv::Mat yuvMat(camera_image_height + camera_image_height / 2, camera_image_width, CV_8UC1, data);
+        cv::cvtColor(yuvMat, img, cv::COLOR_YUV2BGRA_NV21);
+    } else {
+        img = cv::Mat(camera_image_height, camera_image_width, CV_8UC4, data);
     }
-    catch (const std::exception& e) {
-        // If any exception occurs, record it in the output data structure  
-        std::string exceptionMessage = e.what();
-        
-        // Allocate memory for the exception message
-        output.exception = strdup(exceptionMessage.c_str());
-    }
-    // Return the output data structure
-    return output;
+
+    rotateMat(img, rotation);
+    cv::Mat imgNormalized = preprocess_mat(img, model_image_height, model_image_width, mean_values, std_values);
+    return run_inference(index, imgNormalized, model_image_height, model_image_width, objectDetectionFlag, output_data);
 }