+++ ios/Classes/pytorch_ffi.h
@@ -31,3 +31,4 @@ struct ModelLoadResult {
 struct ModelLoadResult load_ml_model(const char* model_path);
 struct OutputData modelInference(int index,float* input_data_ptr);
 struct OutputData image_model_inference(int index, unsigned char* data,int input_length, int height, int width, int objectDetectionFlag, float* mean_values, float* std_values, float* output_data);
+struct OutputData camera_model_inference(int index, unsigned char* data,int rotation,int isYUV, int model_image_height, int model_image_width,int camera_image_height,int camera_image_width, int objectDetectionFlag, float* mean_values, float* std_values, float* output_data);